+++ src/fractal/lib/complex.wgsl
@@ -1,4 +1,4 @@
-fn conj(z: vec2<f64>) -> vec2<f64>
+fn conj(z: vec2<f32>) -> vec2<f32>
 {
     return vec2(
         z.x,
@@ -6,7 +6,7 @@ fn conj(z: vec2<f64>) -> vec2<f64>
     );
 }
 
-fn cis(rot: f64) -> vec2<f64>
+fn cis(rot: f32) -> vec2<f32>
 {
     return vec2(
         cos(rot),
@@ -14,32 +14,32 @@ fn cis(rot: f64) -> vec2<f64>
     );
 }
 
-fn cmul(lhs: vec2<f64>, rhs: vec2<f64>) -> vec2<f64>
+fn cmul(lhs: vec2<f32>, rhs: vec2<f32>) -> vec2<f32>
 {
     return mat2x2(lhs.x, lhs.y, -lhs.y, lhs.x)*rhs;
 }
 
-fn cdiv(lhs: vec2<f64>, rhs: vec2<f64>) -> vec2<f64>
+fn cdiv(lhs: vec2<f32>, rhs: vec2<f32>) -> vec2<f32>
 {
     return mat2x2(lhs.x, lhs.y, lhs.y, -lhs.x)*rhs/norm_sqr(rhs);
 }
 
-fn norm_sqr(x: vec2<f64>) -> f64
+fn norm_sqr(x: vec2<f32>) -> f32
 {
     return dot(x, x);
 }
 
-fn norm(x: vec2<f64>) -> f64
+fn norm(x: vec2<f32>) -> f32
 {
     return sqrt(norm_sqr(x));
 }
 
-fn arg(x: vec2<f64>) -> f64
+fn arg(x: vec2<f32>) -> f32
 {
     return atan2(x.y, x.x);
 }
 
-fn clog(x: vec2<f64>) -> vec2<f64>
+fn clog(x: vec2<f32>) -> vec2<f32>
 {
     return vec2(
         log(norm_sqr(x))/2.0,
@@ -47,12 +47,12 @@ fn clog(x: vec2<f64>) -> vec2<f64>
     );
 }
 
-fn cexp(x: vec2<f64>) -> vec2<f64>
+fn cexp(x: vec2<f32>) -> vec2<f32>
 {
     return exp(x.x)*cis(x.y);
 }
 
-fn powc(x: vec2<f64>, y: vec2<f64>) -> vec2<f64>
+fn powc(x: vec2<f32>, y: vec2<f32>) -> vec2<f32>
 {
     if x.x == 0.0 && x.y == 0.0
     {
